Move modal to separate component, keeping layout as a server component.
